fix(deps): Update dependencies and correct the schema extension
Because of the hard-set version of pydantic, I have to update all of the other dependencies in order to make sure that we get compatible versions.

The extension of ladybug-display-schema was also not correct since it didn't account for the fact that tuples of numbers are used instead of points throughout the ladybug-geometry schemas. For now, it seemed wiser to just remove it rather than having an incorrect and misleading implementation.

I also added an to_polydata method for Polygon2d, which may be needed for certain visualizations.