fix(text alignment): fix the case for upside down text
I'm not really proud of this solution but it will get us over the line until I find a more generic solution to catch this case.

What is happening is that in a very particular case the right hand rule results in a upside-down text. I'm catching the case and flipping the rotation normal.